+++ lib/server.rs
@@ -11,7 +11,7 @@ use std::time::Duration;
 use std::panic;
 
 // Import local files
-use crate::common::reason_phrase;
+use crate::common::{any_string, reason_phrase};
 use crate::handle::handle_connection;
 use crate::header::{headers_to_string, Header};
 use crate::method::Method;
@@ -20,6 +20,8 @@ use crate::request::Request;
 use crate::response::Response;
 use crate::route::Route;
 use crate::VERSION;
+
+pub(crate) type ErrorHandler = dyn Fn(Request, String) -> Response;
 
 /// Defines a server.
 pub struct Server {
@@ -109,7 +111,7 @@ impl Server {
                 Response::new()
                     .status(500)
                     .text(format!("Internal Server Error :/\nError: {}", err))
-                    .header(Header::new("Content-Type", "text/plain"))
+                    .header("Content-Type", "text/plain")
             }),
 
             default_headers: vec![Header::new("Server", format!("afire/{}", VERSION))],
@@ -134,7 +136,7 @@ impl Server {
     ///     Response::new()
     ///         .status(200)
     ///         .text("N O S E")
-    ///         .header(Header::new("Content-Type", "text/plain"))
+    ///         .header("Content-Type", "text/plain")
     /// });
     ///
     /// // Starts the server
@@ -154,8 +156,11 @@ impl Server {
         for event in listener.incoming() {
             // Read stream into buffer
             let mut stream = event.ok()?;
-            stream.set_read_timeout(self.socket_timeout).unwrap();
-            stream.set_write_timeout(self.socket_timeout).unwrap();
+
+            if self.socket_timeout.is_some() {
+                stream.set_read_timeout(self.socket_timeout).unwrap();
+                stream.set_write_timeout(self.socket_timeout).unwrap();
+            }
 
             // Get the response from the handler
             // Uses the most recently defined route that matches the request
@@ -169,12 +174,35 @@ impl Server {
             );
 
             for middleware in &mut self.middleware.iter().rev() {
-                match middleware.borrow_mut().post(req.clone(), res.clone()) {
-                    MiddleResponse::Continue => {}
-                    MiddleResponse::Add(i) => res = i,
-                    MiddleResponse::Send(i) => {
-                        res = i;
-                        break;
+                #[cfg(feature = "panic_handler")]
+                {
+                    let result = panic::catch_unwind(panic::AssertUnwindSafe(|| {
+                        middleware.borrow_mut().post(req.clone(), res.clone())
+                    }));
+
+                    match result {
+                        Ok(i) => match i {
+                            MiddleResponse::Continue => {}
+                            MiddleResponse::Add(i) => res = i,
+                            MiddleResponse::Send(i) => {
+                                res = i;
+                                break;
+                            }
+                        },
+                        Err(e) => res = (self.error_handler)(req.clone(), any_string(e)),
+                    }
+                }
+
+                #[cfg(not(feature = "panic_handler"))]
+                {
+                    let result = middleware.borrow_mut().post(req.clone(), res.clone());
+                    match result {
+                        MiddleResponse::Continue => {}
+                        MiddleResponse::Add(i) => res = i,
+                        MiddleResponse::Send(i) => {
+                            res = i;
+                            break;
+                        }
                     }
                 }
             }
@@ -192,13 +220,11 @@ impl Server {
 
             // Convert the response to a string
             // TODO: Use Bytes instead of String
+            let status = res.status;
             let mut response = format!(
                 "HTTP/1.1 {} {}\r\n{}\r\n\r\n",
-                res.status,
-                match res.reason {
-                    Some(i) => i,
-                    None => reason_phrase(res.status),
-                },
+                status,
+                res.reason.unwrap_or_else(|| reason_phrase(status)),
                 headers_to_string(headers)
             )
             .as_bytes()
@@ -214,30 +240,6 @@ impl Server {
 
         // We should Never Get Here
         None
-    }
-
-    /// Get the ip a server is listening on as a string
-    ///
-    /// For example, "127.0.0.1"
-    /// ## Example
-    /// ```rust
-    /// // Import Library
-    /// use afire::Server;
-    ///
-    /// // Create a server for localhost on port 8080
-    /// let mut server: Server = Server::new("localhost", 8080);
-    ///
-    /// // Get the ip a server is listening on as a string
-    /// assert_eq!("127.0.0.1", server.ip_string());
-    /// ```
-    #[deprecated(since = "0.3.0", note = "Instead use .ip.to_string()")]
-    pub fn ip_string(&self) -> String {
-        self.ip
-            .octets()
-            .iter()
-            .map(|x| x.to_string())
-            .collect::<Vec<String>>()
-            .join(".")
     }
 
     /// Set the satrting buffer size. The default is `1024`
@@ -271,16 +273,20 @@ impl Server {
     /// // Create a server for localhost on port 8080
     /// let mut server: Server = Server::new("localhost", 8080)
     ///     // Add a default header to the response
-    ///     .default_header(Header::new("Content-Type", "text/plain"));
+    ///     .default_header("Content-Type", "text/plain");
     ///
     /// // Start the server
     /// // As always, this is blocking
     /// # server.set_run(false);
     /// server.start().unwrap();
     /// ```
-    pub fn default_header(self, header: Header) -> Server {
+    pub fn default_header<T, K>(self, key: T, value: K) -> Server
+    where
+        T: AsRef<str>,
+        K: AsRef<str>,
+    {
         let mut headers = self.default_headers;
-        headers.push(header);
+        headers.push(Header::new(key.as_ref(), value.as_ref()));
 
         Server {
             default_headers: headers,
@@ -335,6 +341,7 @@ impl Server {
     /// ```
     // I want to change this to be Server builder style
     // But that will require modifying *every* example so that can wait...
+    #[doc(hidden)]
     pub fn set_run(&mut self, run: bool) {
         self.run = run;
     }
@@ -366,157 +373,8 @@ impl Server {
     /// server.start().unwrap();
     /// ```
     #[cfg(feature = "panic_handler")]
-    pub fn error_handler(&mut self, res: fn(Request, String) -> Response) {
+    pub fn error_handler(&mut self, res: impl Fn(Request, String) -> Response + 'static) {
         self.error_handler = Box::new(res);
-    }
-
-    /// Define the panic handler but with closures!
-    ///
-    /// Basicity just [`Server::error_handler`] but with closures
-    ///
-    /// Default response is 500 "Internal Server Error :/"
-    ///
-    /// This is only available if the `panic_handler` feature is enabled
-    ///
-    /// Make sure that this wont panic because then the thread will crash
-    /// ## Example
-    /// ```rust
-    /// // Import Library
-    /// use afire::{Server, Response};
-    ///
-    /// // Create a server for localhost on port 8080
-    /// let mut server: Server = Server::new("localhost", 8080);
-    ///
-    /// // Set the panic handler response
-    /// server.error_handler_c(Box::new(|_req, err| {
-    ///     Response::new()
-    ///         .status(500)
-    ///         .text(format!("Internal Server Error: {}", err))
-    /// }));
-    ///
-    /// // Start the server
-    /// # server.set_run(false);
-    /// server.start().unwrap();
-    /// ```
-    #[cfg(feature = "panic_handler")]
-    pub fn error_handler_c(&mut self, res: Box<dyn Fn(Request, String) -> Response>) {
-        self.error_handler = res;
-    }
-
-    /// Create a new route the runs for all methods and paths
-    ///
-    /// May be useful for a 404 page as the most recently defined route takes priority
-    /// so by defining this route first it would trigger if nothing else matches
-    /// ## Example
-    /// ```rust
-    /// // Import Library
-    /// use afire::{Server, Response, Header, Method};
-    ///
-    /// // Starts a server for localhost on port 8080
-    /// let mut server: Server = Server::new("localhost", 8080);
-    ///
-    /// // Define 404 page
-    /// // Because this is defined first, it will take a low priority
-    /// server.all(|req| {
-    ///     Response::new()
-    ///         .status(404)
-    ///         .text("The page you are looking for does not exist :/")
-    ///         .header(Header::new("Content-Type", "text/plain"))
-    /// });
-    ///
-    /// // Define a route
-    /// // As this is defined last, it will take a high priority
-    /// server.route(Method::GET, "/nose", |req| {
-    ///     Response::new()
-    ///         .status(200)
-    ///         .text("N O S E")
-    ///         .header(Header::new("Content-Type", "text/plain"))
-    /// });
-    ///
-    /// // Starts the server
-    /// // This is blocking
-    /// # server.set_run(false);
-    /// server.start().unwrap();
-    /// ```
-    #[deprecated(since = "0.3.0", note = "Instead use .route(Method::ANY, \"*\", ...)")]
-    pub fn all(&mut self, handler: fn(Request) -> Response) {
-        self.routes
-            .push(Route::new(Method::ANY, "**".to_owned(), Box::new(handler)));
-    }
-
-    /// Create a new route the runs for all methods and paths
-    ///
-    /// May be useful for a 404 page as the most recently defined route takes priority
-    /// so by defining this route first it would trigger if nothing else matches
-    /// ## Example
-    /// ```rust
-    /// // Import Library
-    /// use afire::{Server, Response, Header, Method};
-    ///
-    /// // Starts a server for localhost on port 8080
-    /// let mut server: Server = Server::new("localhost", 8080);
-    ///
-    /// // Define 404 page
-    /// // Because this is defined first, it will take a low priority
-    /// server.all_c(Box::new(|req| {
-    ///     Response::new()
-    ///         .status(404)
-    ///         .text("The page you are looking for does not exist :/")
-    ///         .header(Header::new("Content-Type", "text/plain"))
-    /// }));
-    ///
-    /// // Define a route
-    /// // As this is defined last, it will take a high priority
-    /// server.route(Method::GET, "/nose", |req| {
-    ///     Response::new()
-    ///         .status(200)
-    ///         .text("N O S E")
-    ///         .header(Header::new("Content-Type", "text/plain"))
-    /// });
-    ///
-    /// // Starts the server
-    /// // This is blocking
-    /// # server.set_run(false);
-    /// server.start().unwrap();
-    /// ```
-    #[deprecated(since = "0.3.0", note = "Instead use .route(Method::ANY, \"*\", ...)")]
-    pub fn all_c(&mut self, handler: Box<dyn Fn(Request) -> Response>) {
-        self.routes
-            .push(Route::new(Method::ANY, "**".to_owned(), handler));
-    }
-
-    /// Create a new route for any type of request
-    ///
-    /// Basicity just [`Server::any`] but with closures
-    /// ## Example
-    /// ```rust
-    /// // Import Library
-    /// use afire::{Server, Response, Header};
-    ///
-    /// // Starts a server for localhost on port 8080
-    /// let mut server: Server = Server::new("localhost", 8080);
-    ///
-    /// // Define a route
-    /// server.any("/nose", |req| {
-    ///     Response::new()
-    ///         .status(200)
-    ///         .text("N O S E")
-    ///         .header(Header::new("Content-Type", "text/plain"))
-    /// });
-    ///
-    /// // Starts the server
-    /// // This is blocking
-    /// # server.set_run(false);
-    /// server.start().unwrap();
-    /// ```
-    /// Now you can make any type of request to `/nose` and it will return a 200
-    #[deprecated(since = "0.1.5", note = "Instead use .route(Method::ANY...)")]
-    pub fn any<T>(&mut self, path: T, handler: fn(Request) -> Response)
-    where
-        T: fmt::Display,
-    {
-        self.routes
-            .push(Route::new(Method::ANY, path.to_string(), Box::new(handler)));
     }
 
     /// Create a new route for specified requests
@@ -533,7 +391,7 @@ impl Server {
     ///     Response::new()
     ///         .status(200)
     ///         .text("N O S E")
-    ///         .header(Header::new("Content-Type", "text/plain"))
+    ///         .header("Content-Type", "text/plain")
     /// });
     ///
     /// // Starts the server
@@ -541,49 +399,18 @@ impl Server {
     /// # server.set_run(false);
     /// server.start().unwrap();
     /// ```
-    pub fn route<T>(&mut self, method: Method, path: T, handler: fn(Request) -> Response)
-    where
-        T: fmt::Display,
+    pub fn route<T>(
+        &mut self,
+        method: Method,
+        path: T,
+        handler: impl Fn(Request) -> Response + 'static,
+    ) where
+        T: AsRef<str>,
     {
-        self.routes
-            .push(Route::new(method, path.to_string(), Box::new(handler)));
-    }
-
-    /// Define a new route with a closure as a handler
-    ///
-    /// Basicity just [`Server::route`] but with closures
-    /// ## Example
-    /// ```rust
-    /// // Import Library
-    /// use afire::{Server, Response, Header, Method};
-    ///
-    /// use std::cell::RefCell;
-    ///
-    /// // Create a server for localhost on port 8080
-    /// let mut server: Server = Server::new("localhost", 8080);
-    ///
-    /// let cell = RefCell::new(0);
-    ///
-    /// // Define a route with a closure
-    /// server.route_c(Method::GET, "/nose", Box::new(move |req| {
-    ///     cell.replace_with(|&mut old| old + 1);
-    ///
-    ///     Response::new()
-    ///         .status(200)
-    ///         .text("N O S E")
-    ///         .header(Header::new("Content-Type", "text/plain"))
-    /// }));
-    ///
-    /// // Starts the server
-    /// // This is blocking
-    /// # server.set_run(false);
-    /// server.start().unwrap();
-    /// ```
-    pub fn route_c<T>(&mut self, method: Method, path: T, handler: Box<dyn Fn(Request) -> Response>)
-    where
-        T: fmt::Display,
-    {
-        self.routes
-            .push(Route::new(method, path.to_string(), handler));
+        self.routes.push(Route::new(
+            method,
+            path.as_ref().to_owned(),
+            Box::new(handler),
+        ));
     }
 }