+++ lib/tests/internal/common.rs
@@ -16,3 +16,10 @@ fn decode_url() {
 
     assert_eq!(common::decode_url("%A9".to_owned()), "©");
 }
+
+#[test]
+fn trim_end_bytes() {
+    let mut inp = vec![1, 2, 3, 4, 0, 0, 0, 0];
+    common::trim_end_bytes(&mut inp);
+    assert_eq!(inp, vec![1, 2, 3, 4])
+}